requirements.txt
requests>=2.31.0
PyYAML>=6.0
sqlalchemy>=2.0.0
